requirements.txt
streamlit
langchain-google-genai
python-dotenv
langchain-groq